Improve module augmentation in `minimap` plugin
