+++ packages/ckeditor5-minimap/src/index.ts
@@ -8,3 +8,7 @@
  */
 
 export { default as Minimap } from './minimap';
+
+export type { MinimapConfig } from './minimapconfig';
+
+import './augmentation';
