+++ packages/ckeditor5-minimap/src/minimapconfig.ts
@@ -89,15 +89,3 @@ export interface MinimapConfig {
 	extraClasses?: string;
 
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface EditorConfig {
-
-		/**
-		 * The configuration of the minimap feature. Introduced by the {@link module:minimap/minimap~Minimap} feature.
-		 *
-		 * Read more in {@link module:minimap/minimapconfig~MinimapConfig}.
-		 */
-		minimap?: MinimapConfig;
-	}
-}
